dsp: the new dispatch tests use is_multiple_of; no same-type cast in the 4x4 test
Clippy on the batch: the eight-wide SATD dispatch in both distortion
modules wrote its three divisibility tests with `%`
(manual_is_multiple_of), and the 4x4 transform test cast an i32 to i32
(unnecessary_cast). Distinct clippy warnings: 321 on a2562aa, 324 on
44ea8b8, 317 here; 0 errors throughout. No behaviour change.

+++ src/dsp/hevc_x86_128.rs
@@ -2450,7 +2450,7 @@ mod tests {
                         };
                     }
                 }
-                let bd_shift = 12 - (trial % 3) as i32 * 2;
+                let bd_shift = 12 - (trial % 3) * 2;
                 let mut want = c;
                 (s.idct[0])(&mut want, bd_shift, mx, my);
                 let mut want_dst = c;
